requirements.txt
feedparser>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.28.0
pyyaml>=6.0
